+++ YoutubeDownloader/Services/UpdateService.cs
@@ -6,60 +6,72 @@ using Onova.Services;
 
 namespace YoutubeDownloader.Services
 {
-    public class UpdateService
+    public class UpdateService : IDisposable
     {
         private readonly IUpdateManager _updateManager = new UpdateManager(
             new GithubPackageResolver("Tyrrrz", "YoutubeDownloader", "YoutubeDownloader.zip"),
             new ZipPackageExtractor());
 
-        private Version _updateVersion;
+        private readonly SettingsService _settingsService;
+
+        private Version? _updateVersion;
+        private bool _updatePrepared;
         private bool _updaterLaunched;
 
-        public async Task<Version> CheckPrepareUpdateAsync()
+        public UpdateService(SettingsService settingsService)
         {
+            _settingsService = settingsService;
+        }
+
+        public async Task<Version?> CheckForUpdatesAsync()
+        {
+            if (!_settingsService.IsAutoUpdateEnabled)
+                return null;
+
+            var check = await _updateManager.CheckForUpdatesAsync();
+            return check.CanUpdate ? check.LastVersion : null;
+        }
+
+        public async Task PrepareUpdateAsync(Version version)
+        {
+            if (!_settingsService.IsAutoUpdateEnabled)
+                return;
+
             try
             {
-                // Check for updates
-                var check = await _updateManager.CheckForUpdatesAsync();
-                if (!check.CanUpdate)
-                    return null;
-
-                // Prepare the update
-                await _updateManager.PrepareUpdateAsync(check.LastVersion);
-
-                return _updateVersion = check.LastVersion;
+                await _updateManager.PrepareUpdateAsync(_updateVersion = version);
+                _updatePrepared = true;
             }
             catch (UpdaterAlreadyLaunchedException)
             {
-                return null;
+                // Ignore race conditions
             }
             catch (LockFileNotAcquiredException)
             {
-                return null;
+                // Ignore race conditions
             }
         }
 
         public void FinalizeUpdate(bool needRestart)
         {
+            if (!_settingsService.IsAutoUpdateEnabled)
+                return;
+
+            if (_updateVersion == null || !_updatePrepared || _updaterLaunched)
+                return;
+
             try
             {
-                // Check if an update is pending
-                if (_updateVersion == null)
-                    return;
-
-                // Check if the updater has already been launched
-                if (_updaterLaunched)
-                    return;
-
-                // Launch the updater
                 _updateManager.LaunchUpdater(_updateVersion, needRestart);
                 _updaterLaunched = true;
             }
             catch (UpdaterAlreadyLaunchedException)
             {
+                // Ignore race conditions
             }
             catch (LockFileNotAcquiredException)
             {
+                // Ignore race conditions
             }
         }